requirements.txt

Django==3.1.4
django-ckeditor==5.8.0
django-cleanup==4.0.0
django-crispy-forms==1.8.1
django-js-asset==1.2.2